feat: members / projects 페이지 mainPage에 연결
