+++ src/pages/membersPage/MembersPage.tsx
@@ -1,0 +1,5 @@
+const MembersPage = () => {
+  return <h2>members page</h2>;
+};
+
+export default MembersPage;
