+++ src/shared/constants/routes.ts
@@ -1,4 +1,6 @@
 export const ROUTES = {
   home: '/',
   signUp: '/sign-up',
+  projects: '/projects',
+  members: '/members',
 };
